➖ Removed unused eslint-rules
