+++ lib/gtag.ts
@@ -1,4 +1,3 @@
-/* eslint-disable @typescript-eslint/camelcase */
 export const GA_TRACKING_ID = 'G-NJSH4MRBEQ'
 
 export const pageview = (url: string) => {
